refactor: app bar into separate component
